Standardize the data type for all items in a list
The ansible_facts output can return lists that have items with various
data types. This will break indexing with Elasticsearch since it expects
a single data type for all items in a list. This change iterates through
the list items to set a single standard data type based on the existing
values and types in the list.